Remove edit user from UserPage Change EditModal to EditPostModal. Refactor EditUserModal to have edit password as an option instead for admin page.
