+++ src/components/EditPostModal.tsx
@@ -3,7 +3,7 @@ import { useState, useEffect } from "react";
 import type Post from "../interfaces/Post";
 
 
-export default function EditModal({
+export default function EditPostModal({
   show,
   onHide,
   post,
