+++ src/components/UserInfoCard.tsx
@@ -4,10 +4,9 @@ import type User from '../interfaces/User';
 
 interface Props {
   user: User;
-  handleEditUser: (user: User) => void;
   handleLogout: () => void;
 }
-export default function UserInfoCard({ user, handleEditUser, handleLogout }: Props) {
+export default function UserInfoCard({ user, handleLogout }: Props) {
   return (
     <Col className="border p-4 bg-light">
       <h3>{user.username}</h3>
@@ -17,9 +16,6 @@ export default function UserInfoCard({ user, handleEditUser, handleLogout }: Pro
       <hr />
       <Button variant="danger" onClick={handleLogout}>
         Logout
-      </Button>
-      <Button variant="secondary" className="ms-2" onClick={() => handleEditUser(user)}>
-        Edit information
       </Button>
     </Col>
